fix(core): extract walkTaskGraph util and simplify batch group status
Add walkTaskGraph to task-graph-utils that walks a task graph
topologically using in-degree counters, considering both dependencies
and continuousDependencies. Use it in applyBatchCachedResults.

Simplify TUI batch group status: a batch group is in-progress by
existence (start_batch was called), only completed when all nested
tasks are done.
