+++ packages/nx/src/tasks-runner/task-graph-utils.ts
@@ -274,3 +274,37 @@ function reverseTaskGraph(taskGraph: TaskGraph): TaskGraph {
   }
   return reversed;
 }
+
+/**
+ * Walk a task graph topologically, calling walkFn for each wave of roots.
+ * Considers both dependencies and continuousDependencies.
+ */
+export async function walkTaskGraph(
+  taskGraph: TaskGraph,
+  walkFn: (rootTaskIds: string[]) => Promise<void>
+): Promise<void> {
+  const reversed = reverseTaskGraph(taskGraph);
+  const inDegree: Record<string, number> = {};
+  for (const id of Object.keys(taskGraph.tasks)) {
+    inDegree[id] =
+      (taskGraph.dependencies[id]?.length ?? 0) +
+      (taskGraph.continuousDependencies?.[id]?.length ?? 0);
+  }
+
+  let currentRoots = taskGraph.roots.slice();
+
+  while (currentRoots.length > 0) {
+    await walkFn(currentRoots);
+
+    const nextRoots: string[] = [];
+    for (const id of currentRoots) {
+      for (const depId of reversed.dependencies[id] ?? []) {
+        inDegree[depId]--;
+        if (inDegree[depId] === 0) {
+          nextRoots.push(depId);
+        }
+      }
+    }
+    currentRoots = nextRoots;
+  }
+}
