+++ packages/nx/src/tasks-runner/task-orchestrator.ts
@@ -9,6 +9,7 @@ import { Task, TaskGraph } from '../config/task-graph';
 import { DaemonClient } from '../daemon/client/client';
 import { runCommands } from '../executors/run-commands/run-commands.impl';
 import { getTaskDetails, hashTask, hashTasks } from '../hasher/hash-task';
+import { walkTaskGraph } from './task-graph-utils';
 import { getInputs, TaskHasher } from '../hasher/task-hasher';
 import {
   BatchStatus,
@@ -373,17 +374,8 @@ export class TaskOrchestrator {
 
     const nonCachedTaskIds = new Set<string>();
 
-    // Build in-degree counters and use reverse deps for topological walk
-    const batchReverseDeps = calculateReverseDeps(batch.taskGraph);
-    const inDegree: Record<string, number> = {};
-    for (const id of Object.keys(batch.taskGraph.tasks)) {
-      inDegree[id] = batch.taskGraph.dependencies[id]?.length ?? 0;
-    }
-
-    let currentRoots = batch.taskGraph.roots.slice();
-
-    while (currentRoots.length > 0) {
-      const rootTasks = currentRoots.map((id) => batch.taskGraph.tasks[id]);
+    await walkTaskGraph(batch.taskGraph, async (rootTaskIds) => {
+      const rootTasks = rootTaskIds.map((id) => batch.taskGraph.tasks[id]);
 
       await this.hashBatchTasks(rootTasks);
 
@@ -423,19 +415,7 @@ export class TaskOrchestrator {
           }
         }
       }
-
-      // Find next roots by decrementing dependents' in-degree
-      const nextRoots: string[] = [];
-      for (const id of currentRoots) {
-        for (const depId of batchReverseDeps[id]) {
-          inDegree[depId]--;
-          if (inDegree[depId] === 0) {
-            nextRoots.push(depId);
-          }
-        }
-      }
-      currentRoots = nextRoots;
-    }
+    });
 
     return { cachedResults, needsRehashAfterExecution };
   }
